Split up src/bindinds/mod.rs, add Generator for picking store type
